Add files via upload
Project 1 EC5:
Problem Statement
Let π be a signed permutation of length n. We are interested in analyzing the cycle and component structure of π using a cycle graph representation.
(a) Show that the number of cycles in π can be found in O(n) time. (b) Show that all components in π can be found in O(n) time.