Add files via upload
Problem - Project 1:

Let π be a signed permutation of length 𝑛. We aim to study the cycle and component structure of 𝜋 through its cycle graph representation.

To determine the number of cycles in  𝜋 we construct the cycle graph where vertices represent the elements of π (including their signs), and edges are drawn between elements according to the permutation and adjacency in the signed identity permutation. By traversing this graph and marking visited vertices, each new traversal that begins at an unvisited vertex indicates a new cycle. As each element is visited exactly once, this procedure runs in 𝑂(𝑛) time.

To find all connected components in  π, we again refer to its cycle graph. Each component consists of one or more cycles. Using depth-first or breadth-first traversal starting from unvisited nodes, we can identify all nodes within the same component. Since the graph has 𝑂(𝑛)
vertices and edges, the entire component detection process completes in 
𝑂(𝑛) time.