Add files via upload
Problem - Project 2:

We have a fair coin, meaning it has an equal chance of landing heads or tails on any toss. Our goal is to determine how many times, on average, we need to toss the coin before seeing two heads in a row for the very first time. To find this expected number of tosses, we can approach the problem in two ways. One approach is to carefully analyze the problem using a step-by-step, logical method based on expected values. The other approach is to run numerous simulated coin toss experiments using a Monte Carlo technique to approximate the answer. Here, we focus first on the analytical approach, where we set up and solve equations that represent the different possible outcomes and their impact on the expected number of tosses until the goal is reached.