Add files via upload
Problem – Project 5:

Problem 5.1:
Can you find the parsimony tree using the approximation algorithm for the following matrix M? Is this the maximum parsimony tree? If not, what is the maximum parsimony tree?
Given the matrix M:

    C1	C2	C3	C4
S1	A	B	A	A
S2	A	A	B	A
S3	B	A	A	B
This problem involves computing a minimum spanning tree (MST) using Hamming distances between character strings and comparing it with the maximum parsimony tree, which minimizes the total number of evolutionary changes.

Problem 4.2:
Does the following matrix admit a perfect phylogeny? If not, what is the largest set of compatible characters that does?

Given matrix M:

    C1	C2	C3	C4
S1	0	0	0	1
S2	1	1	0	0
S3	0	0	0	1
S4	0	1	1	0
S5	0	1	1	1

This problem requires analyzing character compatibility using the Four-Gamete Test or subset/disjoint rules on character sets. If any two characters violate compatibility, then the matrix does not admit a perfect phylogeny. The goal is to extract the largest compatible subset.

Problem 4.3:
Given the tree topology for {AC,CT,GT,AT}, can you compute the parsimony length and provide the labeling of internal nodes?
This problem applies the Fitch algorithm, a bottom-up method that identifies minimal changes required at each character position to explain the data. Parsimony length refers to the total number of changes inferred along the tree.

Problem 4.4:
Is there an alternative labeling of internal nodes that gives the same parsimony length as computed in the previous problem?
Even though the tree topology is fixed, there may be multiple internal node labelings that yield the same optimal parsimony score. This problem emphasizes the concept of non-uniqueness in most-parsimonious labelings.